+++ grunt.js
@@ -88,6 +88,7 @@ module.exports = function(grunt) {
         'src/marionette.createObject.js',
         'src/marionette.triggermethod.js',
         'src/marionette.eventbinder.js',
+        'src/marionette.bindEntityEvents.js',
         'src/marionette.eventaggregator.js',
         'src/marionette.controller.js',
         'src/marionette.view.js',