+++ spec/javascripts/view.spec.js
@@ -27,7 +27,7 @@ describe("base view", function(){
     });
   });
 
-  describe("when a view has valid model and collection event configuration", function(){
+  describe("when a view has string-based model and collection event configuration", function(){
     var view;
 
     var View = Backbone.Marionette.View.extend({
@@ -61,6 +61,37 @@ describe("base view", function(){
 
   });
 
+  describe("when a view has function-based model and collection event configuration", function(){
+    var view;
+
+    var View = Backbone.Marionette.View.extend({
+      modelEvents: { 
+        'model-event': jasmine.createSpy("model event handler")
+      },
+      collectionEvents: { 
+        'collection-event': jasmine.createSpy("collection event handler")
+      }
+    });
+
+    beforeEach(function(){
+      view = new View({
+        model: new Backbone.Model(),
+        collection: new Backbone.Collection()
+      });
+    });
+
+    it("should wire up model events", function(){
+      view.model.trigger("model-event");
+      expect(view.modelEvents['model-event']).toHaveBeenCalled();
+    });
+
+    it("should wire up collection events", function(){
+      view.collection.trigger("collection-event");
+      expect(view.collectionEvents['collection-event']).toHaveBeenCalled();
+    });
+
+  });
+
   describe("when a view has model event config with a specified handler method that doesn't exist", function(){
     var getBadViewInstance;
 
@@ -75,7 +106,7 @@ describe("base view", function(){
     });
 
     it("should error when method doesn't exist", function(){
-      expect(getBadViewInstance).toThrow("View method 'does_not_exist' was configured as an event handler, but does not exist.");
+      expect(getBadViewInstance).toThrow("Method 'does_not_exist' was configured as an event handler, but does not exist.");
     });
   });
 });